feat: reduce float decoding allocations by using unsafe
name                           old time/op    new time/op    delta
Decoder_Float64/floats.json-4    28.6µs ± 3%    23.1µs ± 2%   -19.14%  (p=0.000 n=15+14)

name                           old alloc/op   new alloc/op   delta
Decoder_Float64/floats.json-4    2.39kB ± 0%    0.00kB       -100.00%  (p=0.000 n=15+15)

name                           old allocs/op  new allocs/op  delta
Decoder_Float64/floats.json-4       100 ± 0%         0       -100.00%  (p=0.000 n=15+15)
