+++ runtime.go
@@ -1,0 +1,19 @@
+//go:build !appengine && !purego
+
+package jx
+
+import "unsafe"
+
+type sliceType struct {
+	Ptr unsafe.Pointer
+	Len uintptr
+	Cap uintptr
+}
+
+type strType struct {
+	Ptr unsafe.Pointer
+	Len uintptr
+}
+
+//go:linkname noescape runtime.noescape
+func noescape(unsafe.Pointer) unsafe.Pointer
